-- Updated plugin logic to save multiple (instead of just one) pushes (in case multiple ones are received and the app/cordova view is not yet ready and/or has not registered a callback for notifications yet)
E.g.: app is being started up, which may take a couple of seconds, and multiple messages are received during that time...
In that state, app is considered in foreground, so messages need to be temporarily saved until startup is complete (given notification won't be triggered on the tray).
Once the app succesfully registers for notifications, all the pending messages will be pushed, one by one.